Output verify_cli_errors stderr to CI runner
This is to debug the job. It looks like either a caching issue with
Artifacts, or I'm misunderstanding how they should work.

The test is failing on CI - apparently correctly because the output
doesn't include the required message. However the artifact published
for this branch behaves correctly when run locally.

It should be OK to leave the stderr output in, and may be useful
in the future.